# swagger_docs.json
openapi: 3.0.3
info:
  title: 'API Boilerplate'
  description: ''
  version: 1.0.0
servers:
  -
    url: /
    description: ''
paths:
...
  /api/auth/reset-password/check-token:
    ref: 'Reset password check token'
    post:
      operationId: resetPasswordCheckTokenItem
      tags:
        - 'Auth Reset Password'
      responses:
        200:
          description: 'Token checked.'
        401:
          description: Unauthorized.
          content: []
      summary: 'Reset password check token.'
      description: ''
      parameters: []
      requestBody:
        description: 'Reset password check token.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordCheckTokenCredentials'
        required: false
      deprecated: false
    parameters: []
  /api/auth/reset-password/reset:
    ref: 'Reset password'
    post:
      operationId: resetPasswordResetItem
      tags:
        - 'Auth Reset Password'
      responses:
        200:
          description: 'Password reset.'
        401:
          description: Unauthorized.
          content: []
      summary: 'Reset password.'
      description: ''
      parameters: []
      requestBody:
        description: 'Reset password.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordResetCredentials'
        required: false
      deprecated: false
    parameters: []
  /api/auth/reset-password/send-email:
    ref: 'Reset password send email'
    post:
      operationId: resetPasswordSendEmailItem
      tags:
        - 'Auth Reset Password'
      responses:
        200:
          description: 'Email sended.'
        401:
          description: Unauthorized.
          content: []
      summary: 'Reset password send email.'
      description: ''
      parameters: []
      requestBody:
        description: 'Reset password send email.'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordSendEmailCredentials'
        required: false
      deprecated: false
    parameters: []
...
components:
  schemas:
...
    ResetPasswordCheckTokenCredentials:
      type: object
      properties:
        token:
          type: string
          example: string
    ResetPasswordResetCredentials:
      type: object
      properties:
        token:
          type: string
          example: string
        password:
          type: string
          example: string
    ResetPasswordSendEmailCredentials:
      type: object
      properties:
        email:
          type: string
          example: johndoe@example.com
...
  securitySchemes:
    apiKey:
      type: apiKey
      description: 'Value for the Authorization header parameter.'
      name: Authorization
      in: header
security:
  -
    apiKey: []
...